if the format says 'int64', try 'sint' first, try 'int' as fall back (since the type won't be 'sint' unless the value is negative)

+++ src/include/rest_catalog/objects/struct_field.hpp
@@ -30,7 +30,7 @@ public:
 	string TryFromJSON(yyjson_val *obj);
 
 public:
-	uint64_t id;
+	int32_t id;
 	string name;
 	unique_ptr<Type> type;
 	bool required;
